new: Fibers This is the longest I have ever taken to implement a feature, and this time it's the "Fiber", inspired from Bob Nystrom's Wren. The gist of the feature is instead of creating a new stack each time you instantiate a frame, maintain a global stack of values and callframes, and then reuse and if required, reallocate the whole thing as required. This saves a lot of allocation costs in deeply recursive functions, like 'fib' already shows around 50% improvement in benchmarks.
The main caveat of the whole thing is since we are reusing the old stack
forever, we have to be careful to unlink any leftover objects whenever a
frame is popped of, otherwise the ref_decr in store_slot will bite us at
unexpected places whenever that slot is reused for a new frame.

Also, just for the kick of it, I've introduced a new 'incr_ref' bytecode
to increment the refcount of the arguments without loop. Designwise,
this is a pretty shitty design as the bytecode should have absolutely
zero knowledge about the garbage collection mechanism used. This will be
removed soon.