new: operator methods Now, an operator method can be declared inside a class, which will be invoked by the interpreter when that object is used in an arithmetic expression. Operators which can have a method are +, -, *, /, >, <, <=, >=, and, or. This has absolutely zero overhead to usual operator operations as search for the method takes place only when there are no appropiate values for a given operator, the check for which also took place before this anyway. ``` class Test {     pub:         new() {}         op /(x) {             print("Called operator method / with argument ", x, "!\n")         } } obj = Test() z = obj / 4 // should output called operator method / with argument 4 ``` Operator methods are functionally exactly similar to usual object methods and so their return values should depend on the statement written inside the method itself. If no values are returned explicitly, as usual, 'nil' will be returned. Since everybody expects operators to behave in a semantically cohesive way, return the appropiate numeric or boolean values for appropiate operator methods.
Right now operator methods are always left associative, i.e. only (obj
op val) will trigger the operator method 'op' of object obj with
argument val, (val op obj) will not. Also, currently there is no way of
runtime type checking in Next. Going in the future, probably only the
later will be implemented, the former will not.

This commit also fixes the following:
1. The core module is now registered correctly to slot 0 of all other
modules at compilation.
2. For a frame, the instruction pointer is set to the correct position
before throwing an exception, which results in accurate reporting of the
source line.
3. Code generator does not anymore try to optimize imported method
calls, all such calls have been delegated to the runtime as 'call_method' over
the imported module variable itself.
4. 'store_slot' family now increments the ref of the TOP first before
decrementing the slot, so that an object getting assigned to the same
slot doesn't get free'd.
5. 'ret' now also increments the ref of the returned object, to avoid
the case of releasing the object in case there are slots in the previous frame
pointing to the same object.
6. Imported calls to a module using 'call_method' has finally been fixed
to both a) pop the module up, and b) correctly allocate slot for the
object in case the called method is a constructor.