new: arrays Arrays are implemented as a class in the core module. Since no subscript operators are available till now in the language, get and set are implemented as methods in the array class. When the subscript operator arrives, it will be overloaded to use subscript getters and subscript setters.
Arrays are basically manipulated using four builtins :
array_allocate, array_reallocate, array_get and array_set. A new raw
array type Value* is introduced in valuetypes.h, which is only
representable, but not accessible from Next code. array_allocate and
array_reallocate returns the raw array, which is stored as member 'items'
of the array class. Being a collection type, 'items' requires a custom
garbage collection at each 'set' and gc of the array class itself.

The problem is, each get and set operation now trips through a callframe
push -> pritimive call -> callframe pop, which can probably be optimized
if the operations can be implemented either in direct opcode, or in some
form in the interpreter loop itself. This will probably be revisited in
future.