new: better strings Previously, strings were stored in a really horrible way, mapping each hash to a 'string' object directly in a 'HashMap'. Which, among all other inefficiencies, gave absolutely zero protections to hash collisions.
When the new 'map_string' benchmark was added at first, it
was not at all giving the expected 'sum'. Digging a bit, I found that
more than one numbers were mapped to apparently same 'strings' in the
map. This finding was confirmed by manually comparing the affected
string inside a Next program, which had different alphabetical values,
but were returing true on comparison. Finally, I extracted those strings, and
computed their hashes using the StringLibrary, and found out that their
hashes were colliding. Since, the library then just mapped a string to a
hash, and returned true if only the hash already existed in the global
'strings' map, regardless of its actual value, it was returning same hash for
two completely different strings when the collision was occuring.

This commit revamps the string architecture in Next. Now, NextString is
a typedef for _NextString*, which contains the actual string, its length
and its hash. The strings are stored in an 'unordered_set' with a custom
comparator and hasher. The interface to the stringlibrary remains
completely unchanged, except for additions of three new methods:
    1. two 'append' methods join a number of strings
    2. 'get_len' method returns the length of the string
Since NextString is opaque, all other code that previously touched
NextString remain completely unchanged, which was one of the major
goals of this revamp.

Benchmarks are slightly degraded, which I think has to do something with
all the new allocations happening inside 'stringlibrary', we'll look
into that.

After all these work, 'map_string' returns the correct sum now.