new: builtin_module Now, Next provides a streamlined way to define a builtin module, which is initialized when the first time it is imported by a program.
To define a builtin module, add an entry to modules.h containing the
name of the module, and the structure name which contains the init
function, that will be used to register members to the module for the
first time it is imported.
Then, add the module header to modules_includes.h, and we're done.

The init function takes a BuiltinModule*, which contains all the
required functions to add various types of members to the module.

A new random module is defined using the new import system, which also
allows importing of builtin modules now.